engine style refactor
