+++ Cargo.toml
@@ -24,7 +24,8 @@ web-sys = { version = "0.3", features = [
   "HtmlCanvasElement",
   "ImageData",
   "Document",
-  "Window"
+  "Window",
+  "MouseEvent"
 ]}
